Fix bug in Slice arraySize when size is multiple of vector length.
Fix a bug in the Slice `arraySize` function.
When the size is a multiple of the vector length,
the function returns zero for the size of the last index,
which is incorrect; the last index is completely full
and should instead return the vector length.

Also accounts for the case when the size is zero.

Closes #836.